[GUI Components] Split out the gui components into own files
