+++ src/client/gui/widget/button_widget.cpp
@@ -2,7 +2,8 @@
 
 #include <iostream>
 
-#include "../component.h"
+#include "../component/rectangle_component.h"
+#include "../component/text_component.h"
 
 namespace gui {
