+++ src/client/gui/widget/label_widget.cpp
@@ -1,6 +1,6 @@
 #include "../widget.h"
 
-#include "../component.h"
+#include "../component/text_component.h"
 
 namespace gui {
